Several changes in the visualization of the Item Properties to avoid unnecessary copy of objects

+++ src/terralib/layout/item/GridPlanarModel.cpp
@@ -241,6 +241,6 @@ double te::layout::GridPlanarModel::getInitialCoord(double initialCoord, double 
       break;
     }
   }
-  int interval = initialCoord / gap;
+  int interval = (int)(initialCoord / gap);
   return interval * gap;
 }
